feat!: Add option to generate an implemention of `AvroSchema`.
Currently the only option to retrieve a schema for your type is to
enable `derive_schemas` in `GeneratorBuilder`. This will implement
`AvroSchemaComponent` for your type and there is a blanket implementation
`impl<T: AvroSchemaComponent> AvroSchema for T`. However the schema
that is derived does not necessarily match the original schema (even
in canonical form). This is because the naming of fields can change
(capitalized vs lowercase) and probably other edgecases.

This adds an alternative option, which implements `AvroSchema`
directly. It takes the canonical form of the schema used to generate
the type and parses it at runtime. I've chosen to implement `AvroSchema`
instead of `AvroSchemaComponent` as the second one is meant to be
composable, but that would require significantly modifying the schema.

This is a breaking change as the `derive_schemas` option in `GeneratorBuilder`
has been renamed to `implement_avro_schema` and has a different
argument type. The argument type has been changed into an enum, so
that the derive and manual implementation can't both be enabled as
that would result in a compile error (conflicting implementations).
The CLI flag has also been changed to `--impl-schemas` and also
uses the enum.