feat: Allow to override (part of) the generated code for a field (no CLI support)
The primary use case for me is using more specific types directly instead of having
to create adapter structs, for example having `IpAddr` as a field instead of `Vec<u8>`.

I don't have a direct usecase for changing the docstring, but I imagine it can be useful
when you don't control the schema yourself.